refine

+++ vllm/model_executor/models/qwen2_5_vl.py
@@ -1440,9 +1440,9 @@ class Qwen2_5_VLForConditionalGeneration(
         grid_thw_list = grid_thw.tolist()
 
         if image_input["type"] == "image_embeds":
-            image_embeds = image_input["image_embeds"].type(self.visual.dtype)
+            image_embeds = image_input["image_embeds"].to(self.visual.dtype)
         else:
-            pixel_values = image_input["pixel_values"]
+            pixel_values = image_input["pixel_values"].to(self.visual.dtype)
             pixel_values = maybe_do_input_norm(pixel_values, self.input_norm)
 
             if self.use_data_parallel:
@@ -1498,9 +1498,11 @@ class Qwen2_5_VLForConditionalGeneration(
         grid_thw_list = grid_thw.tolist()
 
         if video_input["type"] == "video_embeds":
-            video_embeds = video_input["video_embeds"].type(self.visual.dtype)
+            video_embeds = video_input["video_embeds"].to(self.visual.dtype)
         else:
-            pixel_values_videos = video_input["pixel_values_videos"]
+            pixel_values_videos = video_input["pixel_values_videos"].to(
+                self.visual.dtype
+            )
             pixel_values_videos = maybe_do_input_norm(
                 pixel_values_videos, self.input_norm
             )
